# internal/delivery/http/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/questionnaires": {
            "get": {
                "description": "Возвращает список анкет с фильтрами и пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questionnaires"
                ],
                "summary": "Список анкет",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Фильтр по оплате",
                        "name": "payment",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата с (unix или YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата по (unix или YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.QuestionnairesListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/questionnaires/update": {
            "post": {
                "description": "Обновляет данные анкеты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questionnaires"
                ],
                "summary": "Обновить анкету",
                "parameters": [
                    {
                        "description": "Данные анкеты",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.UpdateQuestionnaireRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Status"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/questionnaires/{id}": {
            "get": {
                "description": "Возвращает анкету по идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questionnaires"
                ],
                "summary": "Получить анкету",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Questionnaire ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.QuestionnaireResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reviews": {
            "get": {
                "description": "Возвращает список отзывов с фильтрами и пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Список отзывов",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по ID пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата с (unix или YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата по (unix или YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ReviewsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "get": {
                "description": "Возвращает отзыв по идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Получить отзыв",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Review ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ReviewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin_history_pkg_proto_gen_go.Answer": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "admin_history_pkg_proto_gen_go.Questionnaire": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin_history_pkg_proto_gen_go.Answer"
                    }
                },
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "history": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "payment": {
                    "type": "boolean"
                },
                "status": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "admin_history_pkg_proto_gen_go.Review": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "server.QuestionnaireResponse": {
            "type": "object",
            "properties": {
                "questionnaire": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.Questionnaire"
                }
            }
        },
        "server.QuestionnairesListResponse": {
            "type": "object",
            "properties": {
                "questionnaires": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin_history_pkg_proto_gen_go.Questionnaire"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "server.ReviewResponse": {
            "type": "object",
            "properties": {
                "review": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.Review"
                }
            }
        },
        "server.ReviewsListResponse": {
            "type": "object",
            "properties": {
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin_history_pkg_proto_gen_go.Review"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "server.Status": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "server.UpdateQuestionnaireRequest": {
            "type": "object",
            "properties": {
                "questionnaire": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.Questionnaire"
                }
            }
        },
...
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. Negative\nsecond values with fractions must still have non-negative nanos values\nthat count forward in time. Must be from 0 to 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch\n1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to\n9999-12-31T23:59:59Z inclusive.",
                    "type": "integer"
                }
            }
        },
...
    }
}